refactored serializer and API client classes
