+++ server/src/serializers/FilmSerializer.js
@@ -24,29 +24,6 @@ class FilmSerializer {
         }) 
         return serializedFilms
     }
-
-    static async getPosters(film) {     
-        let movieId = ""
-        if (film.title === "A New Hope") {
-            movieId = 11
-        } else if (film.title === "The Empire Strikes Back") {
-            movieId = 1891
-        } else if (film.title === "Return of the Jedi") {
-            movieId = 1892
-        } else if (film.title === "The Phantom Menace") {
-            movieId = 1893
-        } else if (film.title === "Attack of the Clones") {
-            movieId = 1894
-        } else if (film.title === "Revenge of the Sith") {
-            movieId = 1895
-        }
-
-        const posterPath = await TheMovieDataBaseClient.getPosterPath(movieId)
-
-        const coverImage = `http://image.tmdb.org/t/p/w342${posterPath}`
-
-        return coverImage
-    }
 }
 
 export default FilmSerializer
